# back/src/main/java/com/bci/tareas/dto/PhoneDTO.java
package com.bci.tareas.dto;

import com.fasterxml.jackson.annotation.JsonInclude;

import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.EqualsAndHashCode;
import lombok.Getter;
import lombok.NoArgsConstructor;
import lombok.Setter;
import lombok.ToString;

@JsonInclude(JsonInclude.Include.NON_DEFAULT)
@Getter
@Setter
@EqualsAndHashCode()
@ToString
@Builder
@NoArgsConstructor
@AllArgsConstructor
public class PhoneDTO {


    private Long id;
    

	private long number;

	private int cityCode;
 
	private String countryCode;


	
	
}